+++ src/http/controller/expense.controller.test.ts
@@ -7,8 +7,8 @@ import TestAgent from 'supertest/lib/agent';
 import { dataSource } from '../../persistence/data-source';
 import { truncateTables } from '../../persistence/helpers';
 import { userRepository } from '../../persistence/repository/user';
-import { authService } from '../../service/auth';
-import { expenseService } from '../../service/expense';
+import { authService } from '../../service/auth.service';
+import { expenseService } from '../../service/expense.service';
 import { signupDtoFactory } from '../../stubs/auth';
 import { factoryMultiplier } from '../../stubs/common';
 import { createExpenseDtoFactory } from '../../stubs/expense';
