+++ src/http/middleware/validation.middleware.test.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { validate } from './validation';
+import { validate } from './validation.middleware';
 import { Request, Response } from 'express';
 
 describe('validation', () => {
